Extract REST integration test payloads into static_test_data module.
Move synthetic and tabular payload fixtures out of common.py so test data is explicit, readable, and matches the Granta MI JSON envelope format.
