+++ tests/rest/test_rest_tabular.py
@@ -41,7 +41,8 @@ from ansys.materials.manager.parsers.rest._rest_reader import (
 )
 from ansys.materials.manager.parsers.rest.rest_material_reader import RestMaterialReader
 
-from .common import minimal_json, tabular_density_section, tabular_elasticity_section
+from .common import minimal_json, tabular_density_section
+from .static_test_data import TABULAR_ELASTICITY_WITH_TEMPERATURE_SECTION
 
 
 class TestTabularReader:
@@ -86,7 +87,7 @@ class TestTabularReader:
         assert density_tq.independent_parameters[0].name == "Temperature"
 
     def test_read_elasticity_isotropic_with_temp(self):
-        section = tabular_elasticity_section([20.0, 100.0], [2e11, 1.9e11], [0.3, 0.31])
+        section = TABULAR_ELASTICITY_WITH_TEMPERATURE_SECTION
         attrs, values = _tabular_reader(
             ("youngs_modulus", "Tensile modulus"),
             ("poissons_ratio", "Poisson's ratio"),
